Filter out insecure cipher suites in TLS configuration
For legacy reasons Go uses by default both secure and insecure ciphers. The insecure are not used unless the Hub has nothing else to offer, which is not the case, but because of various regulations and penetration tests results we need to filter them out.
We however do not want to mess with the crypto package definitions and include or exclude specific ciphers.
Therefore we have to take the list of secure ones from the crypto package and pass them to TLS configuration unmodified.

Signed-off-by: Antonia Trifonova antonia.trifonova@bosch.io